requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
blinker==1.7.0
blosc==1.11.1
boto3==1.34.19
botocore==1.34.19
botocore-stubs==1.34.19
catboost==0.20
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
databricks-cli==0.18.0
dataclasses==0.6
debugpy==1.8.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
dnspython==2.4.2
docker==6.1.3
docutils==0.16
entrypoints==0.4
evidently==0.1.58.dev0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
fastapi-restful==0.5.0
filelock==3.13.1
Flask==3.0.1
fonttools==4.47.2
from-root==1.3.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
graphviz==0.20.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.12.3
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kneed==0.8.5
llvmlite==0.41.1
Mako==1.3.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.9.2
mypy-boto3-s3==1.34.14
mypy-extensions==1.0.0
nest-asyncio==1.6.0
neuro-mf==0.0.5
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
prettytable==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pydantic==1.10.13
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==5.4.1
pyzmq==25.1.2
querystring-parser==1.2.4
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.7.2
s3transfer==0.10.0
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.13.2
shap==0.44.1
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
statsmodels==0.14.1
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
types-awscrt==0.20.0
types-s3transfer==0.10.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.25.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wincertstore==0.2
xgboost==2.0.3
zipp==3.17.0
-e .
